+++ R/plot.CEriskav.R
@@ -70,10 +70,15 @@
 #' \donttest{
 #'    plot(m)
 #' }
-#' ## Alternative options, using ggplot2
+#' # Alternative options, using base
 #' \donttest{
-#'    plot(m, graph = "ggplot2")
+#'    plot(m, graph = "base")
 #' }
+#' # Can also save the plots to an object
+#' p=plot(m)
+#' # And plot them separately
+#' p$eib
+#' p$evi
 #' 
 #' @export
 #'