+++ R/ceaf.plot.R
@@ -1,9 +1,7 @@
 
 #' @rdname ceaf.plot
 #' 
-#' @param graph A string used to select the graphical engine to use for
-#' plotting. Should (partial-) match the two options `"base"` or
-#' `"ggplot2"`. Default value is `"base"`.
+#' @template args-graph
 #' 
 #' @return \item{ceaf}{A ggplot object containing the plot. Returned only if
 #' `graph="ggplot2"`.}
@@ -69,10 +67,10 @@
 #' @export
 #' 
 ceaf.plot.pairwise <- function(mce,
-                               graph = c("base", "ggplot2", "plotly"),
+                               graph = options("bcea.graph"),
                                ...) {
   
-  graph <- match.arg(graph)
+  graph <- unlist(graph)
   
   if (!(requireNamespace("ggplot2", quietly = TRUE) &&
         requireNamespace("grid", quietly = TRUE))) {